+++ modules/ar_agent.c
@@ -1,7 +1,6 @@
 /* Agerun Agent Implementation */
 #include "ar_agent.h"
 #include "ar_method.h"
-#include "ar_methodology.h"
 #include "ar_list.h"
 #include "ar_heap.h"
 #include "ar_map.h"
@@ -60,20 +59,7 @@ ar_agent_t* ar_agent__create_with_method(const ar_method_t *ref_method, const ar
     return own_agent;
 }
 
-ar_agent_t* ar_agent__create(const char *ref_method_name, const char *ref_version, const ar_data_t *ref_context) {
-    if (!ref_method_name) {
-        return NULL;
-    }
-    
-    // Lookup method using global methodology
-    const ar_method_t *ref_method = ar_methodology__get_method(ref_method_name, ref_version);
-    if (!ref_method) {
-        return NULL;  // Method not found
-    }
-    
-    // Delegate to the new function
-    return ar_agent__create_with_method(ref_method, ref_context);
-}
+/* ar_agent__create removed - use ar_agency__create_agent_with_instance instead */
 
 void ar_agent__destroy(ar_agent_t *own_agent) {
     if (!own_agent) {